requirements.txt
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
httpx==0.28.1
python-dotenv==1.0.0


# Google Sheets
gspread==5.10.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1


# OpenAI
openai==0.27.8


# GIS & mapping
shapely==2.0.1
geopandas==0.13.2
folium==0.14.0


# Scheduling & utils
APScheduler==3.10.1.post1
pytz==2023.3
python-dateutil==2.8.2


# Nice-to-have
tqdm==4.65.0